+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- cli: Add event support for command `edit`
 - core: Add api for get the kind of the given id, which can be either an event or a todo.
 
 ## [v0.4.0] - 2025-08-09